+++ tests/test.py
@@ -1,51 +1,53 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-import load_xl
-import os
+from load_xl import load_config, FileParsingError
 
-
-
-# Test envi file parsing and loading into os.environ
+# Test .env file parsing and loading
 try:
-    print("\n[+] Loading envi file...")
-    load_xl.load_envi_file('.envi')
-    print(os.environ['TEST'])
-    print(os.environ['TEST_ENV'])
-except load_xl.EnviFileParsingError as e:
-    print("Error in parsing envi file: {e}")
+    print("\n[+] Loading .env file...")
+    env_config = load_config(".env")
+    print(env_config.get("TEST"))
+    print(env_config.get("TEST_ENV"))
+except FileParsingError as e:
+    print(f"Error in parsing .env file: {e}")
 
 
-# Test config file parsing and loading into os.environ
+# Test .ini file parsing and returning as dict
 try:
-    print("\n[+] Loading config file...")
-    print(load_xl.read_config_file('.configx'))
-    print(os.environ['CONFIG_ENV'])
-except load_xl.ConfigFileParsingError as e:
-    print('Error in parsing config file: {e}')
+    print("\n[+] Loading .ini file...")
+    ini_config = load_config("test.ini")
+    print(ini_config)
+except FileParsingError as e:
+    print(f"Error in parsing .ini file: {e}")
 
-
-# Test ini file parsing and returning as dict
+# Test .yaml file parsing and returning as dict
 try:
-    print("\n[+] Loading ini file...")
-    ini_file = load_xl.load_ini_file('.ini')
-    print(ini_file)
-except load_xl.IniFileParsingError as e:
-    print('Error in parsing ini file: {e}')
+    print("\n[+] Loading .yaml file...")
+    yaml_config = load_config("test.yaml")
+    print(yaml_config)
+except FileParsingError as e:
+    print(f"Error in parsing .yaml file: {e}")
 
-
-# Test yaml file parsing and returning as dict
+# Test .json file parsing and returning as dict
 try:
-    print("\n[+] Loading yaml file...")
-    yaml_file = load_xl.load_yaml_file('.yaml')
-    print(yaml_file)
-except load_xl.YamlFileParsingError as e:
-    print('Error in parsing yaml file: {e}')
+    print("\n[+] Loading .json file...")
+    json_config = load_config("test.json")
+    print(json_config)
+except FileParsingError as e:
+    print(f"Error in parsing .json file: {e}")
 
-
-# Test json file parsing and returning as dict
+# Test .toml file parsing and returning as dict
 try:
-    print("\n[+] Loading json file...")
-    json_data = load_xl.load_json_file('.json')
-    print(json_data)
-except load_xl.JsonFileParsingError as e:
-    print('Error in parsing json file: {e}')
+    print("\n[+] Loading .toml file...")
+    toml_config = load_config("test.toml")
+    print(toml_config)
+except FileParsingError as e:
+    print(f"Error in parsing .toml file: {e}")
+
+# Test .xml file parsing and returning as dict
+try:
+    print("\n[+] Loading .xml file...")
+    xml_config = load_config("test.xml")
+    print(xml_config)
+except FileParsingError as e:
+    print(f"Error in parsing .xml file: {e}")